# backend-rust/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ABP Studio Rust API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/story-chain": {
      "post": {
        "operationId": "create_story_chain",
        "summary": "Migrated from Python create_story_chain",
        "responses": {
          "200": {
            "description": "Compatible JSON response"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/v1/story-fission": {
      "post": {
        "operationId": "create_story_fission",
        "summary": "Migrated from Python create_story_fission",
        "responses": {
          "200": {
            "description": "Compatible JSON response"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/v1/story-fission/{fission_id}/branch/{branch_id}/retry": {
      "post": {
        "operationId": "retry_fission_branch",
        "summary": "Migrated from Python retry_fission_branch",
        "responses": {
          "200": {
            "description": "Compatible JSON response"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/story-fission/{fission_id}/remerge": {
      "post": {
        "operationId": "remerge_fission_story",
        "summary": "Migrated from Python remerge_fission_story",
        "responses": {
          "200": {
            "description": "Compatible JSON response"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  }
}